requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn==1.4.0
xgboost
graphviz
flask
python-dotenv
joblib
pymysql
tensorflow
imblearn
boto3
flask-cors